backend/requirements.txt
fastapi
uvicorn
motor
python-jose
passlib[bcrypt]
python-multipart
bleach
redis
structlog
slowapi
pydantic[email]
pydantic-settings
python-dotenv
argon2-cffi


